[feat] Service 파일 삭제 및 에러 코드 적용 #88
