+++ src/main/java/prefolio/prefolioserver/config/jwt/JwtAuthenticationEntryPoint.java
@@ -1,16 +1,16 @@
 package prefolio.prefolioserver.config.jwt;
 
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.json.JSONObject;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
+import prefolio.prefolioserver.error.ErrorCode;
 
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.time.LocalDateTime;
+
+
+import static prefolio.prefolioserver.error.ErrorResponse.setResponse;
 
 /** 인증되지 않았을 경우(비로그인) AuthenticationEntryPoint 부분에서 AuthenticationException 발생시키면서
  * 비로그인 상태에서 인증 실패 시, AuthenticationEntryPoint로 핸들링되어 이곳에서 처리 */
@@ -23,20 +23,36 @@ public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
             HttpServletResponse response,
             AuthenticationException e
     ) throws IOException {
+        String exception = (String)request.getAttribute("exception");
 
-        // status를 401 에러로 지정
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-        // json 리턴 및 한글 깨짐 수정
-        response.setContentType("application/json;charset=utf-8");
-        JSONObject json = new JSONObject();
-        String errorMsg = "잘못된 접근입니다.";
-        json.put("timestamp", LocalDateTime.now());
-        json.put("code", "[401] UNAUTHORIZED");
-        json.put("message", errorMsg);
-
-        PrintWriter out = response.getWriter();
-        out.print(json);
+        if(exception == null || exception.equals(ErrorCode.NO_TOKEN.getCode())) {
+            setResponse(response, ErrorCode.NO_TOKEN);
+        }
+        //잘못된 타입의 토큰인 경우
+        else if(exception.equals(ErrorCode.INVALID_AUTH_TOKEN.getCode())) {
+            setResponse(response, ErrorCode.INVALID_AUTH_TOKEN);
+        }
+        //DB에 정보가 없는 토큰을 사용했울 때
+        else if(exception.equals(ErrorCode.INVALID_USER_TOKEN.getCode())) {
+            setResponse(response, ErrorCode.INVALID_USER_TOKEN);
+        }
+        //잘못된 서명
+        else if(exception.equals(ErrorCode.INVALID_SIGNATURE.getCode())) {
+            setResponse(response, ErrorCode.INVALID_SIGNATURE);
+        }
+        //토큰 만료된 경우
+        else if(exception.equals(ErrorCode.EXPIRED_TOKEN.getCode())) {
+            setResponse(response, ErrorCode.EXPIRED_TOKEN);
+        }
+        //지원되지 않는 토큰인 경우
+        else if(exception.equals(ErrorCode.UNSUPPORTED_TOKEN.getCode())) {
+            setResponse(response, ErrorCode.UNSUPPORTED_TOKEN);
+        }
+        else {
+            setResponse(response, ErrorCode.FORBIDDEN_USER);
+        }
     }
+
 }
