+++ src/main/java/prefolio/prefolioserver/error/ErrorCode.java
@@ -33,8 +33,8 @@ public enum ErrorCode {
     UNSUPPORTED_TOKEN(UNAUTHORIZED, "AUTH009", "지원하지 않는 토큰입니다"),
     INVALID_SIGNATURE(UNAUTHORIZED, "AUTH010", "잘못된 JWT 서명입니다"),
     /* Kakao */
-    MISMATCH_VERIFICATION_CODE(UNAUTHORIZED, "AUTH011", "인증번호가 일치하지 않습니다"),
-    EXPIRED_VERIFICATION_CODE(UNAUTHORIZED, "AUTH012", "인증번호가 만료되었습니다"),
+    MISMATCH_VERIFICATION_CODE(UNAUTHORIZED, "AUTH011", "인가코드가 일치하지 않습니다"),
+    EXPIRED_VERIFICATION_CODE(UNAUTHORIZED, "AUTH012", "인가코드가 만료되었습니다"),
     INVALID_USER_TOKEN(UNAUTHORIZED, "AUTH013", "서버에 토큰과 일치하는 정보가 없습니다"),
 
     LOGIN_FAILED(UNAUTHORIZED, "AUTH014", "로그인에 실패했습니다"),
